requirements.txt
Flask==1.1.2
numpy==1.20.1
Pillow==8.1.0
